Adding a test of an actual geode class.
Locks of hacks to get RegionJPFTest to work

Overriding JDK classes and hacking geode classes to disable things that
don't work in JPF. Now the RegionJPF test runs and claims there was a
deadlock, but I'm not sure I believe it. For one thing, one of the threads
in the stack dump is only joining for a limited period of time.

Really, this is only useful for small tests.