GEODE-3126: Adding a query message to the protobuf protocol
Adding an oql query request and response to the new protocol.

This works, but we are not returning the field names in the response,
which maybe we should in the case of responses that are structs.

Right now, all possible query results are mapped into a list of structs.
So for example select count(*) from /region would return a list with
a single struct with one field.