Replacing jgroups with a netty based tcp messenger
This is a prototype of a NettyTcpMessenger that replaces the current
JGroupsMessenger. In this PR I've commented out and replaced the old messenger
with this one. It currently creates a single tcp socket to each member it needs
to send a message to and leaves that socket open. On the server side, there is
a single worker thread that processes all messages.

What's missing:
 - backwards compatibility
 - reconnecting and resending messages if the socket fails
 - Handling write failures
 - handling network partitions and auto reconnect (does not return a QuorumChecker)
 - Removing all use of blocking netty methods
 - Lots of testing