GEODE-2153: Post process intermediate query values
Implementing post processing of intermediate values of the query for
CompactRangeIndex and no index queries. Post processing now happens on
values that are returned from region or index iterators while the query
is evaluating, before selections or projections are applied.

Part of this change involves shipping the security principal to remote
query nodes for partitioned regions. We were already shipping the
principal for gfsh queries, so the principal is already required to be
serializable.

*** We still need to test that this messaging change is backwards
compatible.

Adding a test of post processing with queries from a client for
different index types

*** We still need to test the queries from other endpoints are post
processed properly - REST, GFSH, JMX, protobuf