Post processing values in RangeIndexes
PostProcessing in a range index consistents of reevaluating the index
iterators against the current region value and looking for a struct that
matches the expected index key.

If we find one, use that as the post processed struct. If we do not, the
post processing (or a concurrent update) made that struct no longer
match the query, so we do not include that struct in the results.

This change adds a reference to the indexKey in RegionEntryToValuesMap.
However, it also makes that class static so that it no longer has a
reference to the enclosing class, resulting in the same memory overhead.