Adding a ByteSource that checks if it is used by more than one thread
Adding a byte source that checks to see if it is accessed by more than one
thread. That may or may not be an issue, depending on if anything ever uses or
touches the position or limit of this byte source, but it does seem suspicious
if we are using these in multiple threads for PDXInstanceImpl's byte source.