Removing the use of futures in optimizedGet
Preventing the return of shared objects from optimizedGet by completely removed
the code that uses futures to ensure that there is only one get call in
progress at a time.

This is just for debugging, this change would have implications for real
customers with cache loaders.

The behavior, at least for PdxInstances that are returned from this call, may
have changed with 045bb67c7e480bc21a9904c940cc43680a38356b. Before that change,
we were retrieving and copying the returned object in a different way. It's
possible we were getting different PdxInstances in the old code?