Adding a test that demonstrates a PDXInstance is shared
This test shows that two threads that are doing gets at the same time a put
comes in can end up holding onto the same PDXInstance value. This is
problematic because some PdxInstances methods are not thread safe. Although the
underlying bytes are immutatable, the PDXInstance has a ByteSource with a
position field that changes. That means two threads doing serialization or
calling toString on the PdxInstance could result in one or more threads getting
a corrupt read.

It looks like they are ending up with the same instance because of the behavior
in LocalRegion.optimizedGetObject. We use futures to make sure there is only 1
get that goes through, and both threads receive the same value.

Ending up in optimizedGetObject requires a race with the put, because if the
value was in the cache at the beginning of the get it would be returned earlier
in the get process. The test tries to reproduce the case where both get calls
get past that earlier check and end up in optimizedGetObject.