Using a PartitionedRegion.compute to apply changes to a RedisHash for hset
This is a totally hacked up compute method. The idea is to avoid the get
followed by put on the primary, and instead apply a change to the region entry
while it is locked.

AddsDeltaInfo now implements RemoteEntryModification, which is applied while the region
entry is locked, somewhat the way we would apply a delta.

TODO - this is only happening on the primary, I think. I would like to
completely elmininate all of the delta logic and just apply the same
computation on primary and secondary.