GEODE-9223: Fixing the measurement of RedisHash field overhead
The logic to calculate the overhead per field that I made in the previous
change for GEODE-9223 was incorrect. It had two bugs
- It was reusing a byte array for key and value, which the reflection object
sizer will not count twice
- It was not subtracting out the data bytes added from previous entries from
the total overhead.

This caused the constant to change to 52. With some debugging, I see the
overheadPerField ranges from 46 to 56 with this test.